logging_conf renamed to support_functions
logging_conf contain multiple functions which are not related to
logging. Therefore this file was renamed.
