+++ lib/support_functions.py
